AddressRepo supports saving geometry
Added support for saving geometry fields (state plane,
lat/long).  This was removed when back when we decided we
didn't want users to change them using the master address
web application.

However, we need support for saving the information, so that
readdressing can save the geometry from the original
address.  We prevent users from altering them by not showing
the fields in the forms.

Updates #8